refactor(blorktools): rig constraint manager
